+++ tests/nativeapp/test_package_scripts.py
@@ -1,17 +1,14 @@
-from pathlib import Path
 from textwrap import dedent
-from unittest import mock
 from unittest.mock import PropertyMock
 
-import pytest
 from snowcli.cli.nativeapp.manager import (
     InvalidPackageScriptError,
     MissingPackageScriptError,
     NativeAppManager,
 )
+from snowcli.cli.project.definition_manager import DefinitionManager
 from snowflake.connector import ProgrammingError
 
-from tests.project.fixtures import *
 from tests.testing_utils.fixtures import *
 
 NATIVEAPP_MODULE = "snowcli.cli.nativeapp.manager"
@@ -25,6 +22,14 @@ mock_connection = mock.patch(
     "snowcli.cli.common.cli_global_context._CliGlobalContextAccess.connection",
     new_callable=PropertyMock,
 )
+
+
+def _get_na_manager(working_dir):
+    dm = DefinitionManager(working_dir)
+    return NativeAppManager(
+        project_definition=dm.project_definition["native_app"],
+        project_root=dm.project_root,
+    )
 
 
 @mock.patch(NATIVEAPP_MANAGER_EXECUTE_QUERIES)
@@ -47,7 +52,7 @@ def test_package_scripts(
 ):
     mock_conn.return_value = MockConnectionCtx()
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
     native_app_manager._apply_package_scripts()
     assert mock_execute_query.mock_calls == [
         mock.call(expected_call),
@@ -85,7 +90,7 @@ def test_package_scripts(
 @pytest.mark.parametrize("project_definition_files", ["napp_project_1"], indirect=True)
 def test_missing_package_script(mock_execute, project_definition_files):
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
     with pytest.raises(MissingPackageScriptError):
         (working_dir / "002-shared.sql").unlink()
         native_app_manager._apply_package_scripts()
@@ -98,7 +103,7 @@ def test_missing_package_script(mock_execute, project_definition_files):
 @pytest.mark.parametrize("project_definition_files", ["napp_project_1"], indirect=True)
 def test_invalid_package_script(mock_execute, project_definition_files):
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
     with pytest.raises(InvalidPackageScriptError):
         second_file = working_dir / "002-shared.sql"
         second_file.unlink()
@@ -113,7 +118,7 @@ def test_invalid_package_script(mock_execute, project_definition_files):
 @pytest.mark.parametrize("project_definition_files", ["napp_project_1"], indirect=True)
 def test_undefined_var_package_script(mock_execute, project_definition_files):
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
     with pytest.raises(InvalidPackageScriptError):
         second_file = working_dir / "001-shared.sql"
         second_file.unlink()
@@ -141,7 +146,7 @@ def test_package_scripts_w_missing_warehouse_exception(
     )
 
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
 
     with pytest.raises(ProgrammingError) as err:
         native_app_manager._apply_package_scripts()
@@ -163,7 +168,7 @@ def test_package_scripts_w_warehouse_access_exception(
     )
 
     working_dir: Path = project_definition_files[0].parent
-    native_app_manager = NativeAppManager(str(working_dir))
+    native_app_manager = _get_na_manager(str(working_dir))
 
     with pytest.raises(ProgrammingError) as err:
         native_app_manager._apply_package_scripts()